+++ spot/plugins/SPOTMenubar.py
@@ -8,7 +8,13 @@ The ``SPOTMenubar`` plugin provides a default menubar for SPOT.
 ``SPOTMenubar`` is a global plugin.  Only one instance can be opened.
 
 """
+import os.path
 from ginga.rv.plugins import Menubar
+from ginga.gw import Widgets, GwHelp
+
+import spot.icons
+icondir = os.path.split(spot.icons.__file__)[0]
+from spot import __version__
 
 __all__ = ['SPOTMenubar']
 
@@ -29,10 +35,10 @@ class SPOTMenubar(Menubar.Menubar):
         # create a Window pulldown menu, and add it to the menu bar
         wsmenu = menubar.add_name("Workspace")
 
-        item = wsmenu.add_name("Add Workspace")
-        item.add_callback('activated', self.add_workspace_cb)
-        # item = wsmenu.add_name("Delete Workspace")
-        # item.add_callback('activated', self.del_workspace_cb)
+        item = wsmenu.add_name("Open Workspace")
+        item.add_callback('activated', self.open_workspace_cb)
+        item = wsmenu.add_name("Close Workspace")
+        item.add_callback('activated', self.close_workspace_cb)
 
         # # create a Option pulldown menu, and add it to the menu bar
         # optionmenu = menubar.add_name("Option")
@@ -41,25 +47,43 @@ class SPOTMenubar(Menubar.Menubar):
         # plugmenu = menubar.add_name("Plugins")
         # self.w.menu_plug = plugmenu
 
-        # !!TODO!!
-        # # create a Help pulldown menu, and add it to the menu bar
-        # helpmenu = menubar.add_name("Help")
+        # create a Help pulldown menu, and add it to the menu bar
+        helpmenu = menubar.add_name("Help")
 
-        # item = helpmenu.add_name("About")
-        # item.add_callback('activated',
-        #                   lambda *args: self.fv.banner())
+        item = helpmenu.add_name("About")
+        item.add_callback('activated', self.banner)
 
         # item = helpmenu.add_name("Documentation")
         # item.add_callback('activated', lambda *args: self.fv.help())
 
-    def add_workspace_cb(self, w):
-        self.fv.call_global_plugin_method('CPanel', 'new_workspace_cb',
+    def open_workspace_cb(self, w):
+        self.fv.call_global_plugin_method('CPanel', 'open_workspace_cb',
                                           [w], {})
 
-    def del_workspace_cb(self, w):
+    def close_workspace_cb(self, w):
         ws = self.fv.get_current_workspace()
         if ws is not None:
             self.fv.workspace_closed_cb(ws)
+
+    def banner(self, w):
+        # load banner image
+        banner_file = os.path.join(icondir, "spot.svg")
+        img_native = GwHelp.get_image(banner_file, size=(400, 400))
+
+        # create dialog for banner
+        title = f"SPOT v{__version__}"
+        top = Widgets.Dialog(title=title, parent=self.w.menubar,
+                             buttons=[["Close", 0]], modal=False)
+
+        def _close_banner(*args):
+            self.fv.ds.remove_dialog(top)
+
+        top.add_callback('activated', _close_banner)
+        vbox = top.get_content_area()
+        img_w = Widgets.Image(native_image=img_native)
+        vbox.add_widget(img_w, stretch=1)
+
+        self.fv.ds.show_dialog(top)
 
     def __str__(self):
         return 'spotmenubar'